Refactor sqs pollers
